# backend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AreaDetailResponse": {
        "properties": {
          "connected_total": {
            "title": "Connected Total",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "eero_units": {
            "items": {
              "$ref": "#/components/schemas/EeroUnitRow"
            },
            "title": "Eero Units",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "insight_url": {
            "title": "Insight Url",
            "type": "string"
          },
          "is_online": {
            "title": "Is Online",
            "type": "boolean"
          },
          "last_checked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checked"
          },
          "location_name": {
            "title": "Location Name",
            "type": "string"
          },
          "location_type": {
            "enum": [
              "indoor",
              "outdoor"
            ],
            "title": "Location Type",
            "type": "string"
          },
          "network_id": {
            "title": "Network Id",
            "type": "string"
          },
          "network_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network Name"
          },
          "property_id": {
            "title": "Property Id",
            "type": "string"
          },
          "property_name": {
            "title": "Property Name",
            "type": "string"
          },
          "ssid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssid"
          },
          "status": {
            "enum": [
              "online",
              "degraded",
              "offline"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_history": {
            "items": {
              "$ref": "#/components/schemas/StatusHistoryPoint"
            },
            "title": "Status History",
            "type": "array"
          },
          "wan_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wan Ip"
          }
        },
        "required": [
          "id",
          "network_id",
          "location_name",
          "location_type",
          "is_online",
          "status",
          "property_id",
          "property_name",
          "insight_url",
          "eero_units",
          "connected_total",
          "status_history"
        ],
        "title": "AreaDetailResponse",
        "type": "object"
      },
...
      "DashboardResponse": {
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertItem"
            },
            "title": "Alerts",
            "type": "array"
          },
          "available_ssids": {
            "items": {
              "type": "string"
            },
            "title": "Available Ssids",
            "type": "array"
          },
          "avg_latency_ms": {
            "title": "Avg Latency Ms",
            "type": "integer"
          },
          "degraded_count": {
            "title": "Degraded Count",
            "type": "integer"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "heatmap": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Heatmap",
            "type": "array"
          },
          "heatmap_peak": {
            "$ref": "#/components/schemas/HeatCallout"
          },
          "heatmap_quiet": {
            "$ref": "#/components/schemas/HeatCallout"
          },
          "hero_chart": {
            "$ref": "#/components/schemas/DeviceCountsResponse"
          },
          "hst_now": {
            "title": "Hst Now",
            "type": "string"
          },
          "islands": {
            "items": {
              "$ref": "#/components/schemas/IslandSummary"
            },
            "title": "Islands",
            "type": "array"
          },
          "maintenance": {
            "items": {
              "$ref": "#/components/schemas/MaintenanceWindow"
            },
            "title": "Maintenance",
            "type": "array"
          },
          "online_count": {
            "title": "Online Count",
            "type": "integer"
          },
          "outage_count": {
            "title": "Outage Count",
            "type": "integer"
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/PropertyPin"
            },
            "title": "Properties",
            "type": "array"
          },
          "total_devices": {
            "title": "Total Devices",
            "type": "integer"
          },
          "total_networks": {
            "title": "Total Networks",
            "type": "integer"
          },
          "total_properties": {
            "title": "Total Properties",
            "type": "integer"
          }
        },
        "required": [
          "generated_at",
          "hst_now",
          "total_properties",
          "total_networks",
          "total_devices",
          "avg_latency_ms",
          "outage_count",
          "degraded_count",
          "online_count",
          "islands",
          "properties",
          "alerts",
          "hero_chart",
          "heatmap",
          "heatmap_peak",
          "heatmap_quiet"
        ],
        "title": "DashboardResponse",
        "type": "object"
      },
...
      "DeviceCountsResponse": {
        "description": "Contract for `GET /api/v1/properties/{id}/device-counts`.\n\nSame shape is embedded into the dashboard hero chart so the frontend can\nuse a single component for both views.",
        "properties": {
          "series": {
            "items": {
              "$ref": "#/components/schemas/DeviceCountSeries"
            },
            "title": "Series",
            "type": "array"
          },
          "ssid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssid"
          },
          "timestamps": {
            "items": {
              "type": "string"
            },
            "title": "Timestamps",
            "type": "array"
          }
        },
        "required": [
          "timestamps",
          "series"
        ],
        "title": "DeviceCountsResponse",
        "type": "object"
      },
...
      "EeroUnitRow": {
        "properties": {
          "connected_count": {
            "default": 0,
            "title": "Connected Count",
            "type": "integer"
          },
          "firmware_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firmware Version"
          },
          "is_online": {
            "title": "Is Online",
            "type": "boolean"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "location_type": {
            "enum": [
              "indoor",
              "outdoor"
            ],
            "title": "Location Type",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "serial": {
            "title": "Serial",
            "type": "string"
          }
        },
        "required": [
          "serial",
          "location_type",
          "is_online"
        ],
        "title": "EeroUnitRow",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PropertyDetailResponse": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "central_office": {
            "title": "Central Office",
            "type": "string"
          },
          "chart": {
            "$ref": "#/components/schemas/DeviceCountsResponse"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceRow"
            },
            "title": "Devices",
            "type": "array"
          },
          "devices_count": {
            "title": "Devices Count",
            "type": "integer"
          },
          "eero_models": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Eero Models",
            "type": "object"
          },
          "firmware_versions": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Firmware Versions",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "island": {
            "enum": [
              "oahu",
              "maui",
              "big-island",
              "kauai",
              "molokai",
              "lanai"
            ],
            "title": "Island",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "networks": {
            "items": {
              "$ref": "#/components/schemas/NetworkRow"
            },
            "title": "Networks",
            "type": "array"
          },
          "networks_count": {
            "title": "Networks Count",
            "type": "integer"
          },
          "status": {
            "enum": [
              "online",
              "degraded",
              "offline"
            ],
            "title": "Status",
            "type": "string"
          },
          "uptime_pct": {
            "title": "Uptime Pct",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "island",
          "central_office",
          "status",
          "networks_count",
          "devices_count",
          "uptime_pct",
          "chart",
          "networks",
          "devices"
        ],
        "title": "PropertyDetailResponse",
        "type": "object"
      },
...
      "ReportRequest": {
        "description": "Body of `POST /api/v1/properties/{id}/report`.\n\nEmpty `ssids` means \"include all\" \u2014 same UX as the current Django app\n(zero checkboxes \u21d2 all SSIDs).",
        "properties": {
          "ssids": {
            "items": {
              "type": "string"
            },
            "title": "Ssids",
            "type": "array"
          }
        },
        "title": "ReportRequest",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "StatusHistoryPoint": {
        "properties": {
          "checked_at": {
            "title": "Checked At",
            "type": "string"
          },
          "is_online": {
            "title": "Is Online",
            "type": "boolean"
          },
          "response_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Time Ms"
          }
        },
        "required": [
          "checked_at",
          "is_online"
        ],
        "title": "StatusHistoryPoint",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Common Area Monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/areas/{area_id}": {
      "get": {
        "description": "`area_id` is the network_id (e.g. `AKS-001`). Mock-mode short-circuits;\nreal-mode enforces per-property access via the parent property.",
        "operationId": "get_area_api_v1_areas__area_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "area_id",
            "required": true,
            "schema": {
              "title": "Area Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AreaDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Common-area detail (SPEC \u00a75.6)",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/dashboard": {
      "get": {
        "description": "Per-user filtered dashboard.\n\nMock path returns the full set (dev convenience). Real path filters\nproperties to those the caller has access to.",
        "operationId": "get_dashboard_api_v1_dashboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "island",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Island"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 30,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "ssid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ssid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/properties/{property_id}": {
      "get": {
        "description": "Property detail.\n\nMock path uses string IDs (\"aks\", \"prk\", \u2026) so we can't pre-validate\nvia `require_property_access` (which expects int). The DB path enforces\nper-property access inline.",
        "operationId": "get_property_api_v1_properties__property_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Property detail (drives the dashboard drawer)",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/properties/{property_id}/device-counts": {
      "get": {
        "description": "Critical, must-retain feature (SPEC \u00a73 / README backend contract).\n\n`ssid` omitted \u2192 totals (canonical `ssid=\"\"` rows in `connected_device_count`).\n`ssid` present \u2192 filtered series. Same timestamps drive both views.",
        "operationId": "get_device_counts_api_v1_properties__property_id__device_counts_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 30,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "ssid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ssid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCountsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connected-devices time series for stacked-area chart",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/properties/{property_id}/report": {
      "post": {
        "description": "`POST /api/v1/properties/{id}/report` `{ssids: [...]}` \u2192 streams PDF.\n\nEmpty `ssids` means \"include all\" (matches the current Django app).",
        "operationId": "post_property_report_api_v1_properties__property_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate a PDF WiFi network report for the property (SPEC \u00a75.7)",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/properties/{property_id}/ssids": {
      "get": {
        "description": "SPEC \u00a77 \u2014 feeds the SSID dropdown on the Property Detail page.",
        "operationId": "get_property_ssids_api_v1_properties__property_id__ssids_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 30,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Property Ssids Api V1 Properties  Property Id  Ssids Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Distinct SSIDs seen for this property in the window",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/search": {
      "get": {
        "description": "SPEC \u00a75.4 / \u00a78.3 \u2014 global search bar in the header / Cmd+K palette.",
        "operationId": "get_search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Global search across properties, common areas, and network IDs",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}